test(frankenlibc): expand differential-conformance coverage vs glibc
Strengthens the differential test suite that pins FrankenLibc behavior to
glibc, covering the implementation changes in this batch and adjacent
surface:

- scalb SVID semantics (conformance_diff_scalb), fmaxmin family, fromfp,
  llogb, gamma signgam, c23 narrow / minmax math diffs.
- qsort radix / radix16 / i64 fastlane / count8 sort diffs matching the
  new integer radix lane; heapsort diff.
- string/locale: strcasecmp(+simd), strcmp(+simd), regex collating,
  glob tilde, strfmon, strftime/strptime field handling, tokenize fuzz,
  mbstowcs/wcstombs/iconv SIMD diffs, memcpy.
- printf hexfloat/pointer/zt-length, ecvt/fcvt rounding & ndigit0, cvt
  specials, byte-stable ecvt/gcvt, errno-preserved-on-success.
- adds conformance_strtol_family.rs (new tracked conformance test) and
  refreshes the asprintf / c32rtomb / drand48 / cjk-table differential
  and fuzz harnesses.

(Agent-local zz_probe_*/zz_bench_* scratch harnesses are intentionally
left untracked — only curated conformance_* tests are committed.)

+++ crates/frankenlibc-abi/tests/c32rtomb_differential_fuzz.rs
@@ -55,8 +55,8 @@ fn run(f: unsafe extern "C" fn(*mut c_char, u32, *mut c_void) -> usize, c32: u32
     };
     let errno = unsafe { *__errno_location() };
     let ret_i = ret as i64; // (size_t)-1 -> -1
-    let bytes = if ret_i >= 0 && (ret as usize) <= buf.len() {
-        buf[..ret as usize].to_vec()
+    let bytes = if ret_i >= 0 && ret <= buf.len() {
+        buf[..ret].to_vec()
     } else {
         Vec::new()
     };
